requirements.txt
PyYAML>=6
requests>=2
